Add @test_noalloc convenience
For use in unit-testing, @test_noalloc is implemented as a way to check
that a given call is not allocating. To avoid extra dependencies, this
is implemented as a package extension (loaded with Test). Since package
extensions can not themselves export bindings, the macro is implemented
in AllocCheck.jl, but emits an informative error when called without the
extension loaded.