Implement `:func` ABI for calling GPUCompiler emitted code
This ABI is always ccallable, unlike the `:specfunc` ABI which
frequently is not. In the future, it would be good to selectively
re-enable the `:specfunc` ABI for types that are safely ccallable.

The `:specfunc` ABI is interesting to us not just because it's fast.
It allows us to frequently avoid allocations for passing arguments /
return values, unlike the `:func` ABI which (almost) always allocates.